organize types
