+++ packages/portalnetwork/test/networks/lookup.spec.ts
@@ -4,24 +4,19 @@ import { assert, describe, it } from 'vitest'
 import { PortalNetwork } from '../../src/client/index.js'
 import { BaseNetwork, ContentLookup } from '../../src/networks/index.js'
 
-import type { NodeId } from '@chainsafe/enr'
-
-type LookupPeer = {
-  nodeId: NodeId
-  distance: number
-}
-
-const peers: LookupPeer[] = [
-  { nodeId: '0x1', distance: 1 },
-  { nodeId: '0x2', distance: 2 },
-  { nodeId: '0x3', distance: 3 },
-  { nodeId: '0x4', distance: 4 },
-  { nodeId: '0x5', distance: 5 },
-  { nodeId: '0x6', distance: 6 },
-  { nodeId: '0x7', distance: 7 },
-  { nodeId: '0x8', distance: 8 },
-  { nodeId: '0x9', distance: 9 },
-  { nodeId: '0x10', distance: 10 },
+// Cheat since we aren't using the ENR to validate the behavior of the lookup
+// TODO: Convert these to actual randomly generated ENRs
+const peers: any[] = [
+  { enr: '0x1', distance: 1 },
+  { enr: '0x2', distance: 2 },
+  { enr: '0x3', distance: 3 },
+  { enr: '0x4', distance: 4 },
+  { enr: '0x5', distance: 5 },
+  { enr: '0x6', distance: 6 },
+  { enr: '0x7', distance: 7 },
+  { enr: '0x8', distance: 8 },
+  { enr: '0x9', distance: 9 },
+  { enr: '0x10', distance: 10 },
 ]
 
 describe('Custom Peer Heap', () => {
@@ -55,7 +50,7 @@ describe('Custom Peer Heap', () => {
   it('should return top X elements', () => {
     lookup['lookupPeers'].push(peers[4], peers[9], peers[5], peers[8], peers[6], peers[7])
     assert.equal(lookup['lookupPeers'].size(), 9)
-    const top5: LookupPeer[] = []
+    const top5: any[] = []
     while (lookup['lookupPeers'].peek() && top5.length < 5) {
       top5.push(lookup['lookupPeers'].pop()!)
     }
